+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.2.4] - 2023-02-XX
+## [2.3.0] - 2023-04-XX
 
 ### NEW
 - active payment methods are recognized after deactivating and activating the module and activated again correctly
 - in the backend there are new options for control the deleting of unfinished orders
 - New Payment "SEPA" and an alternative CreditCard-Payment as fallback if CreditCard via ACDC is not possible
+- Send Tracking-Information to PayPal
 
 ### Fixed
 - [0007440](https://bugs.oxid-esales.com/view.php?id=7440) Pending orders with oxordernr 0 are deleted before the payment process can be completed
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [0007413](https://bugs.oxid-esales.com/view.php?id=7413) set PPExpress independently of ShippingSets (They will be set later)
 - [0007384](https://bugs.oxid-esales.com/view.php?id=7384) if order declined during 3ds then go back to payment
 - [0007448](https://bugs.oxid-esales.com/view.php?id=7448) show complete value for refund in backend
+- [0007449](https://bugs.oxid-esales.com/view.php?id=7449) discount surcharge can not be < 0
 - [0007450](https://bugs.oxid-esales.com/view.php?id=7450) Mandantory fields are not ignored with creditcard
 
 ## [2.2.3] - 2023-01-26